+++ client/src/components/etf-catalog/Pagination.tsx
@@ -1,0 +1,76 @@
+import { Button } from "@/components/ui/button";
+import { ChevronLeft, ChevronRight } from "lucide-react";
+
+interface PaginationProps {
+  currentPage: number;
+  totalPages: number;
+  onPageChange: (page: number) => void;
+}
+
+export function Pagination({
+  currentPage,
+  totalPages,
+  onPageChange,
+}: PaginationProps) {
+  if (totalPages <= 1) return null;
+
+  return (
+    <div className="mt-8 flex items-center justify-center gap-2">
+      <Button
+        variant="outline"
+        size="sm"
+        onClick={() => onPageChange(Math.max(1, currentPage - 1))}
+        disabled={currentPage === 1}
+      >
+        <ChevronLeft className="mr-1 h-4 w-4" />
+        Previous
+      </Button>
+
+      <div className="flex items-center gap-2">
+        {Array.from({ length: totalPages }, (_, i) => i + 1).map((page) => {
+          // Show first page, last page, current page, and pages around current
+          const showPage =
+            page === 1 ||
+            page === totalPages ||
+            Math.abs(page - currentPage) <= 1;
+
+          const showEllipsis =
+            (page === 2 && currentPage > 3) ||
+            (page === totalPages - 1 && currentPage < totalPages - 2);
+
+          if (!showPage && !showEllipsis) return null;
+
+          if (showEllipsis) {
+            return (
+              <span key={page} className="px-2 text-muted-foreground">
+                ...
+              </span>
+            );
+          }
+
+          return (
+            <Button
+              key={page}
+              variant={currentPage === page ? "default" : "outline"}
+              size="sm"
+              onClick={() => onPageChange(page)}
+              className="w-10"
+            >
+              {page}
+            </Button>
+          );
+        })}
+      </div>
+
+      <Button
+        variant="outline"
+        size="sm"
+        onClick={() => onPageChange(Math.min(totalPages, currentPage + 1))}
+        disabled={currentPage === totalPages}
+      >
+        Next
+        <ChevronRight className="ml-1 h-4 w-4" />
+      </Button>
+    </div>
+  );
+}
