+++ client/src/data/etfs.ts
@@ -1,0 +1,978 @@
+interface ETF {
+  ticker: string;
+  name: string;
+  description: string;
+  assetType: string;
+  category: string;
+  riskLevel: "Low" | "Moderate" | "High";
+  color: string;
+}
+
+interface ETFLiveData {
+  regularMarketPrice?: number;
+  trailingAnnualDividendYield?: number;
+  fiftyTwoWeekChange?: number;
+  expenseRatio?: number;
+}
+
+// Top 100 ETF data - live data fetched separately
+const ETF_DATA: ETF[] = [
+  // Large Cap US Equity
+  {
+    ticker: "VOO",
+    name: "Vanguard S&P 500 ETF",
+    description:
+      "Tracks the S&P 500 index, providing exposure to 500 large-cap U.S. companies.",
+    assetType: "US Equity",
+    category: "Large Blend",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-1))",
+  },
+  {
+    ticker: "IVV",
+    name: "iShares Core S&P 500 ETF",
+    description:
+      "Seeks to track the S&P 500 index representing large-cap U.S. equities.",
+    assetType: "US Equity",
+    category: "Large Blend",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-1))",
+  },
+  {
+    ticker: "SPY",
+    name: "SPDR S&P 500 ETF Trust",
+    description:
+      "The original ETF tracking the S&P 500 index with high liquidity.",
+    assetType: "US Equity",
+    category: "Large Blend",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-1))",
+  },
+  {
+    ticker: "VTI",
+    name: "Vanguard Total Stock Market ETF",
+    description:
+      "Provides broad exposure to the entire U.S. equity market, including small-, mid-, and large-cap stocks.",
+    assetType: "US Equity",
+    category: "Total Market",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-1))",
+  },
+
+  // Growth & Tech
+  {
+    ticker: "QQQ",
+    name: "Invesco QQQ Trust",
+    description:
+      "Tracks the Nasdaq-100 Index, providing exposure to the largest non-financial companies listed on the Nasdaq.",
+    assetType: "US Growth",
+    category: "Large Growth",
+    riskLevel: "High",
+    color: "hsl(var(--chart-5))",
+  },
+  {
+    ticker: "VUG",
+    name: "Vanguard Growth ETF",
+    description:
+      "Seeks to track large-cap U.S. growth stocks with strong growth characteristics.",
+    assetType: "US Growth",
+    category: "Large Growth",
+    riskLevel: "High",
+    color: "hsl(var(--chart-5))",
+  },
+  {
+    ticker: "IWF",
+    name: "iShares Russell 1000 Growth ETF",
+    description:
+      "Tracks the Russell 1000 Growth Index, focusing on large-cap growth stocks.",
+    assetType: "US Growth",
+    category: "Large Growth",
+    riskLevel: "High",
+    color: "hsl(var(--chart-5))",
+  },
+  {
+    ticker: "VGT",
+    name: "Vanguard Information Technology ETF",
+    description: "Provides exposure to technology sector stocks.",
+    assetType: "Technology",
+    category: "Technology",
+    riskLevel: "High",
+    color: "hsl(var(--chart-5))",
+  },
+  {
+    ticker: "XLK",
+    name: "Technology Select Sector SPDR Fund",
+    description: "Tracks the technology sector within the S&P 500.",
+    assetType: "Technology",
+    category: "Technology",
+    riskLevel: "High",
+    color: "hsl(var(--chart-5))",
+  },
+  {
+    ticker: "QQQM",
+    name: "Invesco NASDAQ 100 ETF",
+    description: "Lower-cost version tracking the Nasdaq-100 Index.",
+    assetType: "US Growth",
+    category: "Large Growth",
+    riskLevel: "High",
+    color: "hsl(var(--chart-5))",
+  },
+  {
+    ticker: "SMH",
+    name: "VanEck Semiconductor ETF",
+    description: "Provides exposure to semiconductor companies worldwide.",
+    assetType: "Technology",
+    category: "Technology",
+    riskLevel: "High",
+    color: "hsl(var(--chart-5))",
+  },
+  {
+    ticker: "VONG",
+    name: "Vanguard Russell 1000 Growth ETF",
+    description:
+      "Tracks the Russell 1000 Growth Index for large-cap growth exposure.",
+    assetType: "US Growth",
+    category: "Large Growth",
+    riskLevel: "High",
+    color: "hsl(var(--chart-5))",
+  },
+  {
+    ticker: "MGK",
+    name: "Vanguard Mega Cap Growth ETF",
+    description: "Focuses on mega-cap U.S. growth stocks.",
+    assetType: "US Growth",
+    category: "Large Growth",
+    riskLevel: "High",
+    color: "hsl(var(--chart-5))",
+  },
+  {
+    ticker: "SPYG",
+    name: "SPDR Portfolio S&P 500 Growth ETF",
+    description: "Tracks the S&P 500 Growth Index.",
+    assetType: "US Growth",
+    category: "Large Growth",
+    riskLevel: "High",
+    color: "hsl(var(--chart-5))",
+  },
+  {
+    ticker: "TQQQ",
+    name: "ProShares UltraPro QQQ",
+    description: "3x leveraged ETF tracking the Nasdaq-100 Index.",
+    assetType: "Leveraged",
+    category: "Leveraged",
+    riskLevel: "High",
+    color: "hsl(var(--chart-5))",
+  },
+
+  // International Equity
+  {
+    ticker: "VEA",
+    name: "Vanguard FTSE Developed Markets ETF",
+    description:
+      "Provides exposure to developed market stocks outside the U.S. and Canada.",
+    assetType: "International Equity",
+    category: "Foreign Large Blend",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-2))",
+  },
+  {
+    ticker: "IEFA",
+    name: "iShares Core MSCI EAFE ETF",
+    description:
+      "Tracks the MSCI EAFE Index, providing broad exposure to developed markets outside North America.",
+    assetType: "International Equity",
+    category: "Foreign Large Blend",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-2))",
+  },
+  {
+    ticker: "VXUS",
+    name: "Vanguard Total International Stock ETF",
+    description:
+      "Provides broad exposure to developed and emerging non-U.S. equity markets around the globe.",
+    assetType: "International Equity",
+    category: "Foreign Large Blend",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-2))",
+  },
+  {
+    ticker: "IEMG",
+    name: "iShares Core MSCI Emerging Markets ETF",
+    description:
+      "Tracks the MSCI Emerging Markets Index, providing broad exposure to emerging market stocks.",
+    assetType: "Emerging Markets",
+    category: "Emerging Markets",
+    riskLevel: "High",
+    color: "hsl(var(--chart-2))",
+  },
+  {
+    ticker: "VWO",
+    name: "Vanguard FTSE Emerging Markets ETF",
+    description:
+      "Provides broad exposure to emerging market equities, focusing on large- and mid-cap companies.",
+    assetType: "Emerging Markets",
+    category: "Emerging Markets",
+    riskLevel: "High",
+    color: "hsl(var(--chart-2))",
+  },
+  {
+    ticker: "EFA",
+    name: "iShares MSCI EAFE ETF",
+    description: "Tracks the MSCI EAFE Index for developed market exposure.",
+    assetType: "International Equity",
+    category: "Foreign Large Blend",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-2))",
+  },
+  {
+    ticker: "SCHF",
+    name: "Schwab International Equity ETF",
+    description: "Provides exposure to developed markets outside the U.S.",
+    assetType: "International Equity",
+    category: "Foreign Large Blend",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-2))",
+  },
+  {
+    ticker: "VEU",
+    name: "Vanguard FTSE All-World ex-US ETF",
+    description: "Tracks all international markets excluding the U.S.",
+    assetType: "International Equity",
+    category: "Foreign Large Blend",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-2))",
+  },
+  {
+    ticker: "IXUS",
+    name: "iShares Core MSCI Total Intl Stock ETF",
+    description: "Provides broad international equity exposure.",
+    assetType: "International Equity",
+    category: "Foreign Large Blend",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-2))",
+  },
+  {
+    ticker: "SPDW",
+    name: "SPDR Portfolio Developed World ex-US ETF",
+    description: "Tracks developed markets excluding the U.S.",
+    assetType: "International Equity",
+    category: "Foreign Large Blend",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-2))",
+  },
+  {
+    ticker: "VGK",
+    name: "Vanguard FTSE Europe ETF",
+    description: "Provides exposure to European equity markets.",
+    assetType: "International Equity",
+    category: "Europe Stock",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-2))",
+  },
+  {
+    ticker: "EFV",
+    name: "iShares MSCI EAFE Value ETF",
+    description:
+      "Focuses on value stocks in developed markets outside North America.",
+    assetType: "International Equity",
+    category: "Foreign Large Value",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-2))",
+  },
+  {
+    ticker: "IDEV",
+    name: "iShares Core MSCI Intl Developed Markets ETF",
+    description: "Tracks developed markets outside North America.",
+    assetType: "International Equity",
+    category: "Foreign Large Blend",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-2))",
+  },
+  {
+    ticker: "ACWI",
+    name: "iShares MSCI ACWI ETF",
+    description:
+      "Provides exposure to global equity markets including the U.S.",
+    assetType: "Global Equity",
+    category: "World Stock",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-2))",
+  },
+  {
+    ticker: "VT",
+    name: "Vanguard Total World Stock ETF",
+    description: "Provides exposure to the entire global equity market.",
+    assetType: "Global Equity",
+    category: "World Stock",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-2))",
+  },
+
+  // Value
+  {
+    ticker: "VTV",
+    name: "Vanguard Value ETF",
+    description:
+      "Provides exposure to large-cap U.S. companies with value characteristics, focusing on undervalued stocks.",
+    assetType: "US Equity",
+    category: "Large Value",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-1))",
+  },
+  {
+    ticker: "IVE",
+    name: "iShares S&P 500 Value ETF",
+    description: "Tracks the value segment of the S&P 500.",
+    assetType: "US Equity",
+    category: "Large Value",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-1))",
+  },
+  {
+    ticker: "IWD",
+    name: "iShares Russell 1000 Value ETF",
+    description: "Tracks the Russell 1000 Value Index.",
+    assetType: "US Equity",
+    category: "Large Value",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-1))",
+  },
+  {
+    ticker: "SPYV",
+    name: "SPDR Portfolio S&P 500 Value ETF",
+    description:
+      "Tracks the S&P 500 Value Index, providing exposure to large-cap U.S. value stocks.",
+    assetType: "US Equity",
+    category: "Large Value",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-1))",
+  },
+  {
+    ticker: "VBR",
+    name: "Vanguard Small-Cap Value ETF",
+    description:
+      "Provides exposure to small-cap U.S. companies with value characteristics.",
+    assetType: "US Equity",
+    category: "Small Value",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-1))",
+  },
+
+  // Dividend
+  {
+    ticker: "VIG",
+    name: "Vanguard Dividend Appreciation ETF",
+    description:
+      "Provides exposure to companies with a history of increasing dividends, focusing on quality dividend growth stocks.",
+    assetType: "US Dividend",
+    category: "Large Blend",
+    riskLevel: "Low",
+    color: "hsl(var(--chart-1))",
+  },
+  {
+    ticker: "SCHD",
+    name: "Schwab U.S. Dividend Equity ETF",
+    description:
+      "Focuses on high dividend-yielding U.S. stocks with strong fundamentals.",
+    assetType: "US Dividend",
+    category: "Large Value",
+    riskLevel: "Low",
+    color: "hsl(var(--chart-1))",
+  },
+  {
+    ticker: "VYM",
+    name: "Vanguard High Dividend Yield ETF",
+    description: "Provides exposure to high dividend-yielding U.S. stocks.",
+    assetType: "US Dividend",
+    category: "Large Value",
+    riskLevel: "Low",
+    color: "hsl(var(--chart-1))",
+  },
+  {
+    ticker: "DGRO",
+    name: "iShares Core Dividend Growth ETF",
+    description: "Focuses on dividend growth stocks with strong fundamentals.",
+    assetType: "US Dividend",
+    category: "Large Blend",
+    riskLevel: "Low",
+    color: "hsl(var(--chart-1))",
+  },
+  {
+    ticker: "JEPI",
+    name: "JPMorgan Equity Premium Income ETF",
+    description:
+      "Seeks to deliver monthly income through equity investing and options strategies.",
+    assetType: "US Dividend",
+    category: "Large Blend",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-1))",
+  },
+  {
+    ticker: "JEPQ",
+    name: "JPMorgan Nasdaq Equity Premium Income ETF",
+    description:
+      "Seeks monthly income through Nasdaq investing and options strategies.",
+    assetType: "US Dividend",
+    category: "Large Growth",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-1))",
+  },
+
+  // Bonds
+  {
+    ticker: "BND",
+    name: "Vanguard Total Bond Market ETF",
+    description:
+      "Provides broad exposure to U.S. investment-grade bonds with maturities of more than one year.",
+    assetType: "Bonds",
+    category: "Intermediate Bond",
+    riskLevel: "Low",
+    color: "hsl(var(--chart-3))",
+  },
+  {
+    ticker: "AGG",
+    name: "iShares Core U.S. Aggregate Bond ETF",
+    description:
+      "Tracks the Bloomberg U.S. Aggregate Bond Index, providing broad bond market exposure.",
+    assetType: "Bonds",
+    category: "Intermediate Bond",
+    riskLevel: "Low",
+    color: "hsl(var(--chart-3))",
+  },
+  {
+    ticker: "BNDX",
+    name: "Vanguard Total International Bond ETF",
+    description:
+      "Provides exposure to investment-grade bonds from developed and emerging markets outside the U.S.",
+    assetType: "Bonds",
+    category: "International Bond",
+    riskLevel: "Low",
+    color: "hsl(var(--chart-3))",
+  },
+  {
+    ticker: "TLT",
+    name: "iShares 20+ Year Treasury Bond ETF",
+    description:
+      "Tracks long-term U.S. Treasury bonds with maturities greater than 20 years.",
+    assetType: "Bonds",
+    category: "Long Government",
+    riskLevel: "Low",
+    color: "hsl(var(--chart-3))",
+  },
+  {
+    ticker: "VCIT",
+    name: "Vanguard Intermediate-Term Corp Bond ETF",
+    description: "Provides exposure to intermediate-term U.S. corporate bonds.",
+    assetType: "Bonds",
+    category: "Corporate Bond",
+    riskLevel: "Low",
+    color: "hsl(var(--chart-3))",
+  },
+  {
+    ticker: "MUB",
+    name: "iShares National Muni Bond ETF",
+    description:
+      "Provides exposure to investment-grade municipal bonds from across the United States.",
+    assetType: "Bonds",
+    category: "Muni National",
+    riskLevel: "Low",
+    color: "hsl(var(--chart-3))",
+  },
+  {
+    ticker: "VCSH",
+    name: "Vanguard Short-Term Corporate Bond ETF",
+    description: "Provides exposure to short-term U.S. corporate bonds.",
+    assetType: "Bonds",
+    category: "Short-Term Bond",
+    riskLevel: "Low",
+    color: "hsl(var(--chart-3))",
+  },
+  {
+    ticker: "BSV",
+    name: "Vanguard Short-Term Bond ETF",
+    description: "Tracks short-term U.S. government and corporate bonds.",
+    assetType: "Bonds",
+    category: "Short-Term Bond",
+    riskLevel: "Low",
+    color: "hsl(var(--chart-3))",
+  },
+  {
+    ticker: "IEF",
+    name: "iShares 7-10 Year Treasury Bond ETF",
+    description: "Tracks intermediate-term U.S. Treasury bonds.",
+    assetType: "Bonds",
+    category: "Intermediate Government",
+    riskLevel: "Low",
+    color: "hsl(var(--chart-3))",
+  },
+  {
+    ticker: "VTEB",
+    name: "Vanguard Tax-Exempt Bond ETF",
+    description: "Provides exposure to tax-exempt municipal bonds.",
+    assetType: "Bonds",
+    category: "Muni National",
+    riskLevel: "Low",
+    color: "hsl(var(--chart-3))",
+  },
+  {
+    ticker: "MBB",
+    name: "iShares MBS ETF",
+    description: "Tracks U.S. agency mortgage-backed securities.",
+    assetType: "Bonds",
+    category: "Mortgage-Backed",
+    riskLevel: "Low",
+    color: "hsl(var(--chart-3))",
+  },
+  {
+    ticker: "BIL",
+    name: "SPDR Bloomberg 1-3 Month T-Bill ETF",
+    description: "Provides exposure to short-term U.S. Treasury bills.",
+    assetType: "Bonds",
+    category: "Ultrashort Bond",
+    riskLevel: "Low",
+    color: "hsl(var(--chart-3))",
+  },
+  {
+    ticker: "SGOV",
+    name: "iShares 0-3 Month Treasury Bond ETF",
+    description: "Tracks ultra-short-term U.S. Treasury securities.",
+    assetType: "Bonds",
+    category: "Ultrashort Bond",
+    riskLevel: "Low",
+    color: "hsl(var(--chart-3))",
+  },
+  {
+    ticker: "JPST",
+    name: "JPMorgan Ultra-Short Income ETF",
+    description:
+      "Provides exposure to ultra-short-term investment-grade debt securities.",
+    assetType: "Bonds",
+    category: "Ultrashort Bond",
+    riskLevel: "Low",
+    color: "hsl(var(--chart-3))",
+  },
+  {
+    ticker: "VGIT",
+    name: "Vanguard Intermediate-Term Treasury ETF",
+    description: "Tracks intermediate-term U.S. Treasury bonds.",
+    assetType: "Bonds",
+    category: "Intermediate Government",
+    riskLevel: "Low",
+    color: "hsl(var(--chart-3))",
+  },
+  {
+    ticker: "IUSB",
+    name: "iShares Core Total USD Bond Market ETF",
+    description: "Provides broad exposure to U.S. dollar-denominated bonds.",
+    assetType: "Bonds",
+    category: "Intermediate Bond",
+    riskLevel: "Low",
+    color: "hsl(var(--chart-3))",
+  },
+  {
+    ticker: "LQD",
+    name: "iShares iBoxx Investment Grade Corp Bond ETF",
+    description: "Tracks investment-grade U.S. corporate bonds.",
+    assetType: "Bonds",
+    category: "Corporate Bond",
+    riskLevel: "Low",
+    color: "hsl(var(--chart-3))",
+  },
+  {
+    ticker: "GOVT",
+    name: "iShares U.S. Treasury Bond ETF",
+    description:
+      "Provides exposure to U.S. Treasury bonds across all maturities.",
+    assetType: "Bonds",
+    category: "Intermediate Government",
+    riskLevel: "Low",
+    color: "hsl(var(--chart-3))",
+  },
+  {
+    ticker: "JAAA",
+    name: "Janus Henderson AAA CLO ETF",
+    description: "Invests in AAA-rated collateralized loan obligations.",
+    assetType: "Bonds",
+    category: "Corporate Bond",
+    riskLevel: "Low",
+    color: "hsl(var(--chart-3))",
+  },
+  {
+    ticker: "USHY",
+    name: "iShares Broad USD High Yield Corp Bond ETF",
+    description: "Provides exposure to U.S. high-yield corporate bonds.",
+    assetType: "Bonds",
+    category: "High Yield Bond",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-3))",
+  },
+  {
+    ticker: "BIV",
+    name: "Vanguard Intermediate-Term Bond ETF",
+    description:
+      "Tracks intermediate-term U.S. government and corporate bonds.",
+    assetType: "Bonds",
+    category: "Intermediate Bond",
+    riskLevel: "Low",
+    color: "hsl(var(--chart-3))",
+  },
+  {
+    ticker: "SHY",
+    name: "iShares 1-3 Year Treasury Bond ETF",
+    description: "Tracks short-term U.S. Treasury bonds.",
+    assetType: "Bonds",
+    category: "Short Government",
+    riskLevel: "Low",
+    color: "hsl(var(--chart-3))",
+  },
+  {
+    ticker: "VGSH",
+    name: "Vanguard Short-Term Treasury ETF",
+    description: "Provides exposure to short-term U.S. Treasury securities.",
+    assetType: "Bonds",
+    category: "Short Government",
+    riskLevel: "Low",
+    color: "hsl(var(--chart-3))",
+  },
+
+  // Mid/Small Cap
+  {
+    ticker: "IJH",
+    name: "iShares Core S&P Mid-Cap ETF",
+    description: "Tracks the S&P MidCap 400 Index for mid-cap exposure.",
+    assetType: "US Equity",
+    category: "Mid-Cap Blend",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-1))",
+  },
+  {
+    ticker: "VO",
+    name: "Vanguard Mid-Cap ETF",
+    description: "Provides exposure to mid-cap U.S. stocks.",
+    assetType: "US Equity",
+    category: "Mid-Cap Blend",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-1))",
+  },
+  {
+    ticker: "IJR",
+    name: "iShares Core S&P Small-Cap ETF",
+    description: "Tracks the S&P SmallCap 600 Index.",
+    assetType: "US Equity",
+    category: "Small-Cap Blend",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-1))",
+  },
+  {
+    ticker: "IWM",
+    name: "iShares Russell 2000 ETF",
+    description: "Tracks the Russell 2000 Index for small-cap exposure.",
+    assetType: "US Equity",
+    category: "Small-Cap Blend",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-1))",
+  },
+  {
+    ticker: "VB",
+    name: "Vanguard Small-Cap ETF",
+    description: "Provides broad exposure to small-cap U.S. stocks.",
+    assetType: "US Equity",
+    category: "Small-Cap Blend",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-1))",
+  },
+  {
+    ticker: "IWR",
+    name: "iShares Russell Mid-Cap ETF",
+    description: "Tracks the Russell Midcap Index.",
+    assetType: "US Equity",
+    category: "Mid-Cap Blend",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-1))",
+  },
+  {
+    ticker: "VXF",
+    name: "Vanguard Extended Market ETF",
+    description:
+      "Tracks the S&P Completion Index for small and mid-cap exposure.",
+    assetType: "US Equity",
+    category: "Small-Cap Blend",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-1))",
+  },
+  {
+    ticker: "MDY",
+    name: "SPDR S&P MidCap 400 ETF Trust",
+    description: "Tracks the S&P MidCap 400 Index.",
+    assetType: "US Equity",
+    category: "Mid-Cap Blend",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-1))",
+  },
+
+  // Other Large Cap
+  {
+    ticker: "SPLG",
+    name: "SPDR Portfolio S&P 500 ETF",
+    description: "Low-cost ETF tracking the S&P 500.",
+    assetType: "US Equity",
+    category: "Large Blend",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-1))",
+  },
+  {
+    ticker: "ITOT",
+    name: "iShares Core S&P Total U.S. Stock Market ETF",
+    description: "Provides broad exposure to the entire U.S. stock market.",
+    assetType: "US Equity",
+    category: "Total Market",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-1))",
+  },
+  {
+    ticker: "RSP",
+    name: "Invesco S&P 500 Equal Weight ETF",
+    description: "Equal-weight version of the S&P 500.",
+    assetType: "US Equity",
+    category: "Large Blend",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-1))",
+  },
+  {
+    ticker: "SCHX",
+    name: "Schwab U.S. Large-Cap ETF",
+    description:
+      "Tracks the Dow Jones U.S. Large-Cap Total Stock Market Index.",
+    assetType: "US Equity",
+    category: "Large Blend",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-1))",
+  },
+  {
+    ticker: "VV",
+    name: "Vanguard Large-Cap ETF",
+    description: "Provides exposure to large-cap U.S. stocks.",
+    assetType: "US Equity",
+    category: "Large Blend",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-1))",
+  },
+  {
+    ticker: "IWB",
+    name: "iShares Russell 1000 ETF",
+    description: "Tracks the Russell 1000 Index for large-cap exposure.",
+    assetType: "US Equity",
+    category: "Large Blend",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-1))",
+  },
+  {
+    ticker: "SCHB",
+    name: "Schwab U.S. Broad Market ETF",
+    description: "Provides exposure to the total U.S. stock market.",
+    assetType: "US Equity",
+    category: "Total Market",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-1))",
+  },
+  {
+    ticker: "IVW",
+    name: "iShares S&P 500 Growth ETF",
+    description: "Tracks the growth segment of the S&P 500.",
+    assetType: "US Growth",
+    category: "Large Growth",
+    riskLevel: "High",
+    color: "hsl(var(--chart-5))",
+  },
+  {
+    ticker: "OEF",
+    name: "iShares S&P 100 ETF",
+    description: "Tracks the S&P 100 Index of mega-cap U.S. stocks.",
+    assetType: "US Equity",
+    category: "Large Blend",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-1))",
+  },
+  {
+    ticker: "DIA",
+    name: "SPDR Dow Jones Industrial Average ETF",
+    description: "Tracks the Dow Jones Industrial Average.",
+    assetType: "US Equity",
+    category: "Large Blend",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-1))",
+  },
+  {
+    ticker: "QUAL",
+    name: "iShares MSCI USA Quality Factor ETF",
+    description: "Focuses on U.S. stocks with strong quality characteristics.",
+    assetType: "US Equity",
+    category: "Large Blend",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-1))",
+  },
+  {
+    ticker: "DFAC",
+    name: "Dimensional U.S. Core Equity 2 ETF",
+    description: "Provides diversified U.S. equity exposure with factor tilts.",
+    assetType: "US Equity",
+    category: "Large Blend",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-1))",
+  },
+  {
+    ticker: "IUSG",
+    name: "iShares Core S&P U.S. Growth ETF",
+    description: "Tracks the S&P 900 Growth Index.",
+    assetType: "US Growth",
+    category: "Large Growth",
+    riskLevel: "High",
+    color: "hsl(var(--chart-5))",
+  },
+  {
+    ticker: "USMV",
+    name: "iShares MSCI USA Min Vol Factor ETF",
+    description: "Focuses on U.S. stocks with lower volatility.",
+    assetType: "US Equity",
+    category: "Large Blend",
+    riskLevel: "Low",
+    color: "hsl(var(--chart-1))",
+  },
+  {
+    ticker: "SCHG",
+    name: "Schwab U.S. Large-Cap Growth ETF",
+    description: "Tracks large-cap U.S. growth stocks.",
+    assetType: "US Growth",
+    category: "Large Growth",
+    riskLevel: "High",
+    color: "hsl(var(--chart-5))",
+  },
+  {
+    ticker: "DYNF",
+    name: "BlackRock U.S. Equity Factor Rotation ETF",
+    description:
+      "Provides factor-based U.S. equity exposure with dynamic rotation.",
+    assetType: "US Equity",
+    category: "Large Blend",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-1))",
+  },
+
+  // Sector ETFs
+  {
+    ticker: "XLF",
+    name: "Financial Select Sector SPDR Fund",
+    description: "Tracks the financial sector within the S&P 500.",
+    assetType: "Financials",
+    category: "Financials",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-4))",
+  },
+  {
+    ticker: "XLV",
+    name: "Health Care Select Sector SPDR Fund",
+    description: "Tracks the health care sector within the S&P 500.",
+    assetType: "Healthcare",
+    category: "Healthcare",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-4))",
+  },
+  {
+    ticker: "XLE",
+    name: "Energy Select Sector SPDR Fund",
+    description: "Tracks the energy sector within the S&P 500.",
+    assetType: "Energy",
+    category: "Energy",
+    riskLevel: "High",
+    color: "hsl(var(--chart-4))",
+  },
+  {
+    ticker: "XLC",
+    name: "Communication Services Select Sector SPDR Fund",
+    description: "Tracks the communication services sector within the S&P 500.",
+    assetType: "Communication",
+    category: "Communication",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-4))",
+  },
+  {
+    ticker: "XLY",
+    name: "Consumer Discretionary Select Sector SPDR Fund",
+    description: "Tracks the consumer discretionary sector within the S&P 500.",
+    assetType: "Consumer Discretionary",
+    category: "Consumer Cyclical",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-4))",
+  },
+  {
+    ticker: "XLI",
+    name: "Industrial Select Sector SPDR Fund",
+    description: "Tracks the industrial sector within the S&P 500.",
+    assetType: "Industrials",
+    category: "Industrials",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-4))",
+  },
+
+  // Real Estate
+  {
+    ticker: "VNQ",
+    name: "Vanguard Real Estate ETF",
+    description:
+      "Provides broad exposure to U.S. real estate investment trusts (REITs) and real estate companies.",
+    assetType: "REIT",
+    category: "Real Estate",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-4))",
+  },
+
+  // Commodities & Crypto
+  {
+    ticker: "GLD",
+    name: "SPDR Gold Trust",
+    description: "Provides exposure to physical gold bullion.",
+    assetType: "Commodities",
+    category: "Commodities",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-6))",
+  },
+  {
+    ticker: "IAU",
+    name: "iShares Gold Trust",
+    description: "Tracks the price of physical gold.",
+    assetType: "Commodities",
+    category: "Commodities",
+    riskLevel: "Moderate",
+    color: "hsl(var(--chart-6))",
+  },
+  {
+    ticker: "SLV",
+    name: "iShares Silver Trust",
+    description: "Provides exposure to physical silver bullion.",
+    assetType: "Commodities",
+    category: "Commodities",
+    riskLevel: "High",
+    color: "hsl(var(--chart-6))",
+  },
+  {
+    ticker: "IBIT",
+    name: "iShares Bitcoin Trust",
+    description: "Provides exposure to Bitcoin through an ETF structure.",
+    assetType: "Crypto",
+    category: "Cryptocurrency",
+    riskLevel: "High",
+    color: "hsl(var(--chart-6))",
+  },
+  {
+    ticker: "FBTC",
+    name: "Fidelity Wise Origin Bitcoin Fund",
+    description: "Provides exposure to Bitcoin.",
+    assetType: "Crypto",
+    category: "Cryptocurrency",
+    riskLevel: "High",
+    color: "hsl(var(--chart-6))",
+  },
+];
+
+export type { ETF, ETFLiveData };
+export { ETF_DATA };
